requirements.txt
absl-py==0.6.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.7.1
binwalk==2.1.1
Brlapi==0.6.6
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.8.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cvxopt==1.2.3
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.0.6
dnspython==1.15.0
dnspython3==1.15.0
Flask==1.0.2
future==0.16.0
gast==0.2.0
grpcio==1.16.0
gTTS==2.0.3
gTTS-token==1.1.3
h5py==2.8.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-autotime==0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
joblib==0.13.2
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
knockpy==4.1.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
line-bot-sdk==1.8.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.1.1
namedlist==1.7
netifaces==0.10.4
networkx==2.2
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.0.25
pandas==0.24.2
parso==0.3.3
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
ply==3.11
prompt-toolkit==2.0.8
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
Pygments==2.3.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyqtgraph==0.10.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.2
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
redis==3.2.1
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
soupsieve==1.9.2
SQLAlchemy==1.2.10
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.23
usb-creator==0.3.3
virtualenv==16.0.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wpull==2.0.1
xattr==0.9.2
xkit==0.0.0
Yapsy==1.11.223
zope.interface==4.3.2
